requirements.txt
fastapi
uvicorn
python-dotenv
groq
AppOpener
pywhatkit
beautifulsoup4
pillow
rich
requests
keyboard
cohere
googlesearch-python
mtranslate
edge-tts
webdriver-manager
aiofiles
httpx